+++ RiskOfChaos/ModifierController/SkillSlots/SkillSlotModificationData.cs
@@ -1,4 +1,5 @@
-﻿using RoR2;
+﻿using RiskOfChaos.Utilities;
+using RoR2;
 
 namespace RiskOfChaos.ModifierController.SkillSlots
 {
@@ -10,10 +11,16 @@ namespace RiskOfChaos.ModifierController.SkillSlots
         public bool ForceActivate = false;
 
         public float CooldownScale = 1f;
+        public sbyte StockAdds = 0;
 
         public SkillSlotModificationData(SkillSlot slotIndex)
         {
             SlotIndex = slotIndex;
+        }
+
+        public void AddStocksSafe(int diff)
+        {
+            StockAdds = ClampedConversion.Int8(StockAdds + diff);
         }
     }
 }